requirements.txt
ultralytics
patool